Add and update files via upload
Enhanced Physics class, now objects collide almost perfectly
Display has now a Player attribute and uses the Control class, the player attribute if  used by the Control object
Controllable has now two methods, for pressing and releasing a button

+++ src/main/java/io/github/engine/DisplayRefresh.java
@@ -33,7 +33,7 @@ public final class DisplayRefresh implements Runnable {
 	/**
 	 * @return {@code deltaTime}
 	 */
-	public static double getDeltaTime(){
+	public static float getDeltaTime(){
 		return deltaTime;
 	}
